Django/crud_project/requirements.txt
Django==4.1.7
mysqlclient==2.1.1
djangorestframework==3.14.0